+++ mlir/test/Quantum/ResolveGateLevelAdjointTest.mlir
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// RUN: quantum-opt --adjoint-lowering-canonicalization %s | FileCheck %s
+// RUN: quantum-opt --resolve-gate-level-adjoint --split-input-file %s | FileCheck %s
 
 // CHECK-LABEL: test_hermitian_adjoint_canonicalize
 func.func @test_hermitian_adjoint_canonicalize() -> !quantum.bit {
@@ -20,10 +20,12 @@ func.func @test_hermitian_adjoint_canonicalize() -> !quantum.bit {
     %1 = quantum.extract %0[ 0] : !quantum.reg -> !quantum.bit
     // CHECK: [[reg:%.+]] = quantum.alloc( 1) : !quantum.reg
     // CHECK: [[qubit:%.+]] = quantum.extract [[reg]][ 0] : !quantum.reg -> !quantum.bit
-    %2 = quantum.custom "Hadamard"() %1 {adjoint}: !quantum.bit
+    %2 = quantum.custom "Hadamard"() %1 adj : !quantum.bit
     // CHECK:  quantum.custom "Hadamard"() [[qubit]] : !quantum.bit
     return %2 : !quantum.bit
 }
+
+// -----
 
 // CHECK-LABEL: test_rotation_adjoint_canonicalize
 func.func @test_rotation_adjoint_canonicalize(%arg0: f64) -> !quantum.bit {
@@ -31,11 +33,13 @@ func.func @test_rotation_adjoint_canonicalize(%arg0: f64) -> !quantum.bit {
     %1 = quantum.extract %0[ 0] : !quantum.reg -> !quantum.bit
     // CHECK: [[reg:%.+]] = quantum.alloc( 1) : !quantum.reg
     // CHECK: [[qubit:%.+]] = quantum.extract [[reg]][ 0] : !quantum.reg -> !quantum.bit
-    %2 = quantum.custom "RX"(%arg0) %1 {adjoint}: !quantum.bit
+    %2 = quantum.custom "RX"(%arg0) %1 adj : !quantum.bit
     // CHECK: [[arg0neg:%.+]] = arith.negf %arg0 : f64
     // CHECK:  quantum.custom "RX"([[arg0neg]]) [[qubit]] : !quantum.bit
     return %2 : !quantum.bit
 }
+
+// -----
 
 // CHECK-LABEL: test_multirz_adjoint_canonicalize
 func.func @test_multirz_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !quantum.bit) {
@@ -48,9 +52,11 @@ func.func @test_multirz_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !quan
 
     // CHECK: [[arg0neg:%.+]] = arith.negf %arg0 : f64
     // CHECK: [[ret:%.+]]:2 = quantum.multirz([[arg0neg]]) [[qubit1]], [[qubit2]] : !quantum.bit, !quantum.bit
-    %3:2 = quantum.multirz (%arg0) %1, %2 {adjoint} : !quantum.bit, !quantum.bit
+    %3:2 = quantum.multirz (%arg0) %1, %2 adj  : !quantum.bit, !quantum.bit
     return %3#0, %3#1 : !quantum.bit, !quantum.bit
 }
+
+// -----
 
 // CHECK-LABEL: test_pcphase_adjoint_canonicalize
 func.func @test_pcphase_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !quantum.bit) {
@@ -63,9 +69,11 @@ func.func @test_pcphase_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !quan
 
     // CHECK: [[arg0neg:%.+]] = arith.negf %arg0 : f64
     // CHECK: [[ret:%.+]]:2 = quantum.pcphase([[arg0neg]], dim : 2) [[qubit1]], [[qubit2]] : !quantum.bit, !quantum.bit
-    %3:2 = quantum.pcphase(%arg0, dim : 2) %1, %2 {adjoint} : !quantum.bit, !quantum.bit
+    %3:2 = quantum.pcphase(%arg0, dim : 2) %1, %2 adj  : !quantum.bit, !quantum.bit
     return %3#0, %3#1 : !quantum.bit, !quantum.bit
 }
+
+// -----
 
 // CHECK-LABEL: test_paulirot_adjoint_canonicalize
 func.func @test_paulirot_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !quantum.bit, !quantum.bit) {
@@ -80,14 +88,16 @@ func.func @test_paulirot_adjoint_canonicalize(%arg0: f64) -> (!quantum.bit, !qua
 
     // CHECK: [[arg0neg:%.+]] = arith.negf %arg0 : f64
     // CHECK: [[ret:%.+]]:3 = quantum.paulirot ["X", "Y", "Z"]([[arg0neg]]) [[qubit1]], [[qubit2]], [[qubit3]] : !quantum.bit, !quantum.bit, !quantum.bit
-    %4:3 = quantum.paulirot ["X", "Y", "Z"](%arg0) %1, %2, %3 {adjoint} : !quantum.bit, !quantum.bit, !quantum.bit
+    %4:3 = quantum.paulirot ["X", "Y", "Z"](%arg0) %1, %2, %3 adj  : !quantum.bit, !quantum.bit, !quantum.bit
     return %4#0, %4#1, %4#2 : !quantum.bit, !quantum.bit, !quantum.bit
 }
+
+// -----
 
 // CHECK-LABEL: test_gphase_adjoint_canonicalize
 func.func @test_gphase_adjoint_canonicalize(%arg0: f64) {
     // CHECK: [[arg0neg:%.+]] = arith.negf %arg0 : f64
     // CHECK: quantum.gphase([[arg0neg]])
-    quantum.gphase(%arg0) {adjoint}
+    quantum.gphase(%arg0) adj 
     return
 }
